component directories
