+++ src/components/Deprecated/FxBloom.vue
