+++ src/components/Deprecated/FxGradient.vue
