+++ src/components/Deprecated/FxShine.vue
