+++ src/components/Filters/FxBlur.vue
@@ -1,9 +1,9 @@
 <script lang="ts" setup>
-import type { FxEmits } from '../lib/emits'
+import type { FxEmits } from '../../lib/emits'
 import { inject, onBeforeUnmount, onUnmounted } from 'vue'
-import { makeReactive } from '../lib/componentBuilder'
-import commonProps from '../lib/props'
-import { InjectionKeys } from '../lib/provider'
+import { makeReactive } from '../../lib/componentBuilder'
+import commonProps from '../../lib/props'
+import { InjectionKeys } from '../../lib/provider'
 
 const props = defineProps({
   external: {
@@ -11,12 +11,11 @@ const props = defineProps({
     default: false,
   },
   color: commonProps.color,
+  quality: commonProps.quality,
   x: commonProps.x,
   y: commonProps.y,
-  decay: commonProps.decay,
-  power: commonProps.power,
-  samples: commonProps.samples,
-  intensity: commonProps.intensity,
+  steps: commonProps.steps,
+  strength: commonProps.strength,
 })
 const emit = defineEmits<FxEmits>()
 
@@ -26,16 +25,15 @@ const fxController = props.external ? gameObject.filters?.external : gameObject.
 if (!fxController) {
   throw new Error(`filters.${props.external ? 'external' : 'internal'} is not available. Make sure the game object supports filters and WebGL renderer is enabled.`)
 }
-const filter = fxController.addShadow(props.x, props.y, props.decay, props.power, props.color, props.samples, props.intensity)
+const filter = fxController.addBlur(props.quality, props.x, props.y, props.strength, props.color, props.steps)
 
 makeReactive(row => [
-  row('color', () => props.color!, (v: number) => filter.color = v),
+  row('quality', () => props.quality!, (v: number) => filter.quality = v),
   row('x', () => props.x!, (v: number) => filter.x = v),
   row('y', () => props.y!, (v: number) => filter.y = v),
-  row('decay', () => props.decay!, (v: number) => filter.decay = v),
-  row('power', () => props.power!, (v: number) => filter.power = v),
-  row('samples', () => props.samples!, (v: number) => filter.samples = v),
-  row('intensity', () => props.intensity!, (v: number) => filter.intensity = v),
+  row('steps', () => props.steps!, (v: number) => filter.steps = v),
+  row('strength', () => props.strength!, (v: number) => filter.strength = v),
+  row('color', () => props.color!, (v: number) => filter.color = v),
 ])
 
 emit('create', filter)
