+++ src/components/GameObjects/Circle.vue
@@ -1,29 +1,29 @@
 <script lang="ts" setup>
-import type { GameObjectEmits } from '../lib/emits'
+import type { GameObjectEmits } from '../../lib/emits'
 import * as Phaser from 'phaser'
 import { inject } from 'vue'
-import { defineGameObject, makeGameObjectReactive, makeReactive } from '../lib/componentBuilder'
-import commonProps, { gameObjectProps } from '../lib/props'
-import { InjectionKeys } from '../lib/provider'
-import setters from '../lib/setters'
+import { defineGameObject, makeGameObjectReactive, makeReactive } from '../../lib/componentBuilder'
+import commonProps, { gameObjectProps } from '../../lib/props'
+import { InjectionKeys } from '../../lib/provider'
+import setters from '../../lib/setters'
 
 const props = defineProps({
   ...gameObjectProps,
-  points: commonProps.points,
+  radius: commonProps.radius,
   fillColor: commonProps.fillColor,
   fillAlpha: commonProps.fillAlpha,
   lineWidth: commonProps.lineWidth,
   strokeColor: commonProps.strokeColor,
   strokeAlpha: commonProps.strokeAlpha,
 })
-defineEmits<GameObjectEmits<Phaser.GameObjects.Polygon>>()
+defineEmits<GameObjectEmits<Phaser.GameObjects.Arc>>()
 
 const scene = inject(InjectionKeys.Scene)!
-const object = new Phaser.GameObjects.Polygon(scene, props.x || 0, props.y || 0, props.points)
+const object = new Phaser.GameObjects.Arc(scene, props.x || 0, props.y || 0, props.radius)
 
 makeGameObjectReactive(props, object)
 makeReactive(row => [
-  row('points', () => props.points!, setters.points(object)),
+  row('radius', () => props.radius!, setters.radius(object)),
   row('fillColor', () => props.fillColor!, setters.fillColor(object)),
   row('fillAlpha', () => props.fillAlpha!, setters.fillAlpha(object)),
   row('lineWidth', () => props.lineWidth!, setters.lineWidth(object)),
