+++ src/components/GameObjects/Sprite.vue
@@ -1,12 +1,12 @@
 <script lang="ts">
 import type { PropType } from 'vue'
-import type { SpriteEmits } from '../lib/emits'
+import type { SpriteEmits } from '../../lib/emits'
 import * as Phaser from 'phaser'
 import { getCurrentInstance, inject } from 'vue'
-import { defineGameObject, makeGameObjectReactive, makeReactive } from '../lib/componentBuilder'
-import commonProps, { gameObjectProps } from '../lib/props'
-import { InjectionKeys } from '../lib/provider'
-import setters from '../lib/setters'
+import { defineGameObject, makeGameObjectReactive, makeReactive } from '../../lib/componentBuilder'
+import commonProps, { gameObjectProps } from '../../lib/props'
+import { InjectionKeys } from '../../lib/provider'
+import setters from '../../lib/setters'
 
 const SPRITE_EMITS = [
   { attr: 'onAnimationstart', emit: 'animationstart' },
