+++ src/components/Physics/StaticBody.vue
@@ -1,11 +1,11 @@
 <script lang="ts" setup>
 import type * as Phaser from 'phaser'
-import type { BodyEmits } from '../lib/emits'
+import type { BodyEmits } from '../../lib/emits'
 import { inject, onBeforeUnmount } from 'vue'
-import { makeReactive } from '../lib/componentBuilder'
-import commonProps from '../lib/props'
-import { InjectionKeys } from '../lib/provider'
-import setters from '../lib/setters'
+import { makeReactive } from '../../lib/componentBuilder'
+import commonProps from '../../lib/props'
+import { InjectionKeys } from '../../lib/provider'
+import setters from '../../lib/setters'
 
 const props = defineProps({
   width: commonProps.width,
