+++ src/index.ts
@@ -1,36 +1,36 @@
-export { default as Body } from './components/Body.vue'
-export { default as Circle } from './components/Circle.vue'
-export { default as Container } from './components/Container.vue'
-export { default as FxBarrel } from './components/FxBarrel.vue'
-export { default as FxBloom } from './components/FxBloom.vue'
-export { default as FxBlur } from './components/FxBlur.vue'
-export { default as FxBokeh } from './components/FxBokeh.vue'
-export { default as FxCircle } from './components/FxCircle.vue'
-export { default as FxColorMatrix } from './components/FxColorMatrix.vue'
-export { default as FxDisplacement } from './components/FxDisplacement.vue'
-export { default as FxGlow } from './components/FxGlow.vue'
-export { default as FxGradient } from './components/FxGradient.vue'
-export { default as FxPixelate } from './components/FxPixelate.vue'
-export { default as FxShadow } from './components/FxShadow.vue'
-export { default as FxShine } from './components/FxShine.vue'
-export { default as FxVignette } from './components/FxVignette.vue'
-export { default as FxWipe } from './components/FxWipe.vue'
+export { default as FxBloom } from './components/Deprecated/FxBloom.vue'
+export { default as FxCircle } from './components/Deprecated/FxCircle.vue'
+export { default as FxGradient } from './components/Deprecated/FxGradient.vue'
+export { default as FxShine } from './components/Deprecated/FxShine.vue'
+export { default as RoundRectangle } from './components/Deprecated/RoundRectangle.vue'
+export { default as FxBarrel } from './components/Filters/FxBarrel.vue'
+export { default as FxBlur } from './components/Filters/FxBlur.vue'
+export { default as FxBokeh } from './components/Filters/FxBokeh.vue'
+export { default as FxColorMatrix } from './components/Filters/FxColorMatrix.vue'
+export { default as FxDisplacement } from './components/Filters/FxDisplacement.vue'
+export { default as FxGlow } from './components/Filters/FxGlow.vue'
+export { default as FxPixelate } from './components/Filters/FxPixelate.vue'
+export { default as FxShadow } from './components/Filters/FxShadow.vue'
+export { default as FxVignette } from './components/Filters/FxVignette.vue'
+export { default as FxWipe } from './components/Filters/FxWipe.vue'
 export { default as Game } from './components/Game.vue'
-export { default as Image } from './components/Image.vue'
-export { default as Light } from './components/Light.vue'
-export { default as Line } from './components/Line.vue'
-export { default as NineSlice } from './components/NineSlice.vue'
-export { default as Polygon } from './components/Polygon.vue'
-export { default as Rectangle } from './components/Rectangle.vue'
-export { default as RenderTexture } from './components/RenderTexture.vue'
-export { default as RoundRectangle } from './components/RoundRectangle.vue'
+export { default as Circle } from './components/GameObjects/Circle.vue'
+export { default as Container } from './components/GameObjects/Container.vue'
+export { default as Image } from './components/GameObjects/Image.vue'
+export { default as Light } from './components/GameObjects/Light.vue'
+export { default as Line } from './components/GameObjects/Line.vue'
+export { default as NineSlice } from './components/GameObjects/NineSlice.vue'
+export { default as Polygon } from './components/GameObjects/Polygon.vue'
+export { default as Rectangle } from './components/GameObjects/Rectangle.vue'
+export { default as RenderTexture } from './components/GameObjects/RenderTexture.vue'
+export { default as Sprite } from './components/GameObjects/Sprite.vue'
+export { default as Text } from './components/GameObjects/Text.vue'
+export { default as TilemapLayer } from './components/GameObjects/TilemapLayer.vue'
+export { default as Triangle } from './components/GameObjects/Triangle.vue'
+export { default as Zone } from './components/GameObjects/Zone.vue'
+export { default as Body } from './components/Physics/Body.vue'
+export { default as StaticBody } from './components/Physics/StaticBody.vue'
 export { default as Scene } from './components/Scene.vue'
-export { default as Sprite } from './components/Sprite.vue'
-export { default as StaticBody } from './components/StaticBody.vue'
-export { default as Text } from './components/Text.vue'
-export { default as TilemapLayer } from './components/TilemapLayer.vue'
-export { default as Triangle } from './components/Triangle.vue'
-export { default as Zone } from './components/Zone.vue'
 export { onPostUpdate, onPreUpdate, refPhaserInstance, useGame, useScene } from './lib/utils'
 export type { TimelineConfig, TweenConfig } from './types'
 export * as Phavuer from './types'
